+++ src/main/java/com/irontechspace/dynamicdq/executor/task/model/Task.java
@@ -1,25 +1,25 @@
-package com.irontechspace.dynamicdq.rabbit.model;
+package com.irontechspace.dynamicdq.executor.task.model;
 
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
+import lombok.*;
 
 import java.util.List;
 import java.util.UUID;
 
 @Getter
 @Setter
+@Builder
 @ToString
-public class RabbitTask {
+@NoArgsConstructor
+@AllArgsConstructor
+public class Task {
 
     // Ид задачи для логирования и возврата результата пользователю
     private UUID id;
-
     private UUID userId;
     private List<String> userRoles;
+    private List<TaskConfig> configs;
 
-    private List<RabbitTaskConfig> configs;
-
+    private TaskConfigEvents events;
 
     // Рекурсивная задача после выполнения текущей
 //    private String resultPath;