+++ src/main/java/com/irontechspace/dynamicdq/rabbit/model/RabbitTaskOutputQueue.java
@@ -1,4 +1,0 @@
-package com.irontechspace.dynamicdq.rabbit.model;
-
-public class RabbitTaskOutputQueue {
-}