+++ src/main/java/com/irontechspace/dynamicdq/rabbit/model/RabbitTaskOutputType.java
@@ -1,3 +1,0 @@
-package com.irontechspace.dynamicdq.rabbit.model;
-
-public enum RabbitTaskOutputType { database, queue }